requirements.txt
asgiref==3.7.2
attrs==25.3.0
black==25.1.0
click==8.1.8
Django==5.0.1
django-cors-headers==4.3.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.2
drf-spectacular==0.28.0
exceptiongroup==1.2.2
flake8==7.1.2
freezegun==1.5.1
gunicorn==22.0.0
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.3.6
pluggy==1.5.0
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.8.0
pytest==8.3.5
pytest-django==4.10.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytils==0.4.1
pytz==2023.4
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
six==1.17.0
sqlparse==0.4.4
tomli==2.2.1
typing_extensions==4.9.0
uritemplate==4.2.0
